moved storage hook into utils
